remove methods from Dictionary && move tests && add functions to index.ts
