+++ src/dict/Dictionary.ts
@@ -11,8 +11,6 @@ import { beginCell, Builder } from "../boc/Builder";
 import { Cell } from "../boc/Cell";
 import { Slice } from "../boc/Slice";
 import { Maybe } from "../utils/maybe";
-import { generateMerkleProof } from "./generateMerkleProof";
-import { generateMerkleUpdate } from "./generateMerkleUpdate";
 import { parseDict } from "./parseDict";
 import { serializeDict } from "./serializeDict";
 import { deserializeInternalKey, serializeInternalKey } from "./utils/internalKeySerializer";
@@ -365,14 +363,6 @@ export class Dictionary<K extends DictionaryKeyTypes, V> {
 
         // Store
         serializeDict(prepared, resolvedKey.bits, resolvedValue.serialize, builder);
-    }
-
-    generateMerkleProof(key: K): Cell {
-        return generateMerkleProof(this, key)
-    }
-
-    generateMerkleUpdate(key: K, newValue: V): Cell {
-        return generateMerkleUpdate(this, key, newValue);
     }
 }
